ensure float precision

+++ l5kit/l5kit/evaluation/csv_utils.py
@@ -159,9 +159,9 @@ def write_pred_csv(
     assert num_modes <= MAX_MODES
 
     # generate always a fixed size json for MAX_MODES by padding the arrays with zeros
-    coords_padded = np.zeros((num_example, MAX_MODES, future_len, num_coords))
+    coords_padded = np.zeros((num_example, MAX_MODES, future_len, num_coords), dtype=coords.dtype)
     coords_padded[:, :num_modes] = coords
-    confs_padded = np.zeros((num_example, MAX_MODES))
+    confs_padded = np.zeros((num_example, MAX_MODES), dtype=confs.dtype)
     confs_padded[:, :num_modes] = confs
     # TODO check sum to 1
     coords_keys_list = [_generate_coords_keys(future_len, mode_index=idx) for idx in range(MAX_MODES)]
